add mutable test cases

+++ src/main/java/org/sportradar/scoreboard/services/ScoreBoardService.java
@@ -3,15 +3,15 @@ package org.sportradar.scoreboard.services;
 import org.sportradar.scoreboard.ScoreBoardDAO;
 import org.sportradar.scoreboard.entities.Match;
 import org.sportradar.scoreboard.exceptions.DuplicateMatchException;
-import org.sportradar.scoreboard.exceptions.InvalidInputException;
 import org.sportradar.scoreboard.exceptions.MatchNotFoundException;
-import org.sportradar.scoreboard.services.ScoreBoardService;
+import org.sportradar.scoreboard.exceptions.NotAllowedException;
 
 import java.util.List;
 
 /**
  * A service keeping the score board and the operation on that for starting or updating a match
  * and showing a summary of score board
+ *
  * @author Hesam.Karimian
  * @since 12.11.2023
  */
@@ -25,6 +25,7 @@ public class ScoreBoardService {
 
     /**
      * Get summary of the ongoing matches in an ordering way.
+     *
      * @return An ordered list of {@link Match}
      */
     public List<Match> getSummary() {
@@ -33,9 +34,10 @@ public class ScoreBoardService {
 
     /**
      * Start a new match and save it to the score board
+     *
      * @param homeTeamName: name of the team that plays at home.
      * @param awayTeamName: name of the team that plays away.
-     * @throws InvalidInputException if teamName is null or blank
+     * @throws IllegalArgumentException if teamName is null or blank
      */
     public Integer startNewMatch(String homeTeamName, String awayTeamName) {
         validateTeamName(homeTeamName);
@@ -49,13 +51,14 @@ public class ScoreBoardService {
 
     /**
      * Updates the score board with new scores.
-     * @param matchId id of the match.
+     *
+     * @param matchId       id of the match.
      * @param homeTeamScore number of goals home team had already scored.
      * @param awayTeamScore number of goals away team had already scored.
      * @throws IllegalArgumentException if teamNames are null or blank
      * @throws IllegalArgumentException if new scores are 0-0 or negative values.
      * @throws IllegalArgumentException if new scores are lower than previous values.
-     * @throws MatchNotFoundException if mach does not exist on the board with given HomeTeam and AwayTeam.
+     * @throws MatchNotFoundException   if mach does not exist on the board with given HomeTeam and AwayTeam.
      */
     public void updateScore(Integer matchId, int homeTeamScore, int awayTeamScore) {
         if (matchId == null || matchId < 0) {
@@ -74,12 +77,16 @@ public class ScoreBoardService {
 
     /**
      * Finishes a match and remove it from the score board.
+     *
      * @param matchId id of the match to be finished.
      * @throws MatchNotFoundException if mach does not exist on the board with given HomeTeam and AwayTeam.
      */
     public void finishMatch(Integer matchId) {
-    Match match = findMatch(matchId);
-    scoreBoardDAO.delete(match);
+        if (matchId == null || matchId <= 0) {
+            throw new IllegalArgumentException("Id cannot be null");
+        }
+        Match match = findMatch(matchId);
+        scoreBoardDAO.delete(match);
     }
 
     private static void validateDifferentTeamNames(String homeTeamName, String awayTeamName) {
@@ -90,7 +97,7 @@ public class ScoreBoardService {
 
     private static void validateTeamName(String teamName) {
         if (teamName == null || teamName.isBlank()) {
-            throw new InvalidInputException("Team Name");
+            throw new IllegalArgumentException("Team Name should not be null or empty.");
         }
     }
 
@@ -100,8 +107,17 @@ public class ScoreBoardService {
             throw new DuplicateMatchException(match.getHomeTeam().getName(), match.getAwayTeam().getName());
         });
         scoreBoardDAO.findByMatch(getReverseMatch(match)).ifPresent(match1 -> {
-            throw new DuplicateMatchException(match.getHomeTeam().getName(), match.getAwayTeam().getName());
+            throw new DuplicateMatchException(match.getAwayTeam().getName(), match.getHomeTeam().getName());
         });
+
+        scoreBoardDAO.findByTeam(match.getHomeTeam().getName())
+                .ifPresent(m -> {
+                    throw new NotAllowedException("Adding new match", match.getHomeTeam().getName() + " is already playing with another team");
+                });
+        scoreBoardDAO.findByTeam(match.getAwayTeam().getName())
+                .ifPresent(m -> {
+                    throw new NotAllowedException("Adding new match", match.getHomeTeam().getName() + " is already playing with another team");
+                });
     }
 
     private static Match getReverseMatch(Match match) {
